requirements.txt
# Benchmark suite dependencies (001-llm-benchmark-suite)
PyYAML>=6.0
python-Levenshtein>=0.20.0
requests>=2.31.0
numpy>=1.24.0
python-dotenv>=1.0.0
